requirements.txt
python-telegram-bot==20.0
instaloader
ffmpeg-python
requests
